feat(login): update files to load the new login with animation
