+++ Screens/Login.tsx
@@ -1,15 +1,16 @@
-import React, { useEffect, FC, useState } from "react";
+import React, { useEffect, FC, useState, useRef } from "react";
 import * as WebBrowser from "expo-web-browser";
 import * as Google from "expo-auth-session/providers/google";
-import { StyleSheet, Text, View, Image, TouchableOpacity } from "react-native";
+import { StyleSheet, Text, View, Image, TouchableOpacity, Animated, Easing } from "react-native";
 import karotBunny from "../Images/karot-bunny-logo.png";
 import karotSlogan from "../Images/karot-slogan.png";
-import {
-	useFonts,
-	Sarabun_700Bold,
-	Sarabun_400Regular,
-	Sarabun_300Light,
-} from "@expo-google-fonts/sarabun";
+import logoCombinedImage from '../Images/logo/logo_combined.png';
+// import {
+// 	useFonts,
+// 	Sarabun_700Bold,
+// 	Sarabun_400Regular,
+// 	Sarabun_300Light,
+// } from "@expo-google-fonts/sarabun";
 import { LinearGradient } from "expo-linear-gradient";
 import { LoginPropsType, GlobalStateType } from "../Utils/types";
 import { useSelector, useDispatch } from "react-redux";
@@ -21,11 +22,12 @@ const Login: FC<LoginPropsType> = ({
 	loggedIn,
 	setLoggedIn,
 }: LoginPropsType) => {
-	let [fontsLoaded] = useFonts({
-		Sarabun_700Bold,
-		Sarabun_400Regular,
-		Sarabun_300Light,
-	});
+	// let [fontsLoaded] = useFonts({
+	// 	Sarabun_700Bold,
+	// 	Sarabun_400Regular,
+	// 	Sarabun_300Light,
+	// });
+  const splashImagesAnimation = useRef(new Animated.Value(0)).current;
 	const userData = useSelector<GlobalStateType>((state) => state.user.data);
 	const dispatch = useDispatch();
 	const [accessToken, setAccessToken] = useState<string | undefined>();
@@ -76,6 +78,23 @@ const Login: FC<LoginPropsType> = ({
 		});
 	}
 
+	useEffect(() => {
+    const playSplashAnimation = async (): Promise<void> => {
+      // app is ready, hide SplashScreen, start animation
+      
+      Animated.timing(splashImagesAnimation, {
+        delay: 2000,
+        toValue: -100,
+        duration: 2000,
+        useNativeDriver: true,
+        easing: Easing.bezier(0.65, 0, 0.35, 1)
+      }).start(() => setLoggedIn({ status: '', screen: 'login' }));
+    }
+		playSplashAnimation();
+  }, []);
+
+	
+
 	return (
 		<View style={styles.container}>
 			<LinearGradient
@@ -84,8 +103,27 @@ const Login: FC<LoginPropsType> = ({
 				colors={["#2383C9", "#5A1E70"]}
 				style={styles.container}
 			>
-				<Image source={karotBunny} style={styles.bunnyLogo} />
-				<Image source={karotSlogan} style={styles.karotSlogan} />
+				<Animated.View style={[styles.splashImageContainer,
+        {
+          transform: [
+            { translateY: splashImagesAnimation }
+          ]
+        }
+      ]}
+      >
+        <Image 
+          style={{
+            maxWidth: 215,
+            width: "50%",
+            height: 352,
+          }}
+          resizeMode="contain"
+          source={logoCombinedImage}
+          fadeDuration={0}
+        />
+      </Animated.View>
+			{/* <Image source={logoCombinedImage} style={styles.bunnyLogo} />
+				<Image source={karotSlogan} style={styles.karotSlogan} /> */}
 				<TouchableOpacity
 					disabled={!request}
 					onPress={() => {
@@ -131,6 +169,11 @@ const styles = StyleSheet.create({
 		fontSize: 20,
 		fontFamily: "Sarabun_700Bold",
 	},
+  splashImageContainer: {
+    width: '100%',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
 });
 
 export default Login;
